Updated app.js to match IMiddlewareMap hook.
The original app.js bot.use had "dialog" as the middleware hook in the example usage of "bot.use". According to IMiddlewareMap's definition "dialog" is not a valid hook and thus causes the "Logging" section to never run.

The function defined for "dialog" has "session" and "next" as parameters which corresponds to IMiddlewareMap's "botbuilder" hook.

I updated app.js to change "dialog" to "botbuilder".

The example app.js now runs without error and outputs the logging messages.